backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
marshmallow
python-dotenv
psycopg[binary]
